requirement.txt
django==3.2
django-mdeditor==0.1.20